+++ src/client/components/Swiper/index.js
@@ -32,6 +32,10 @@ import { rootDirectory } from '@uixkit.react/config';
 
 export default class Swiper extends Component {
 
+	constructor(props) {
+		super(props);
+	}
+	
 	componentDidMount(){
 		
 		//