+++ src/client/views/_pages/ComponentsDemo/index.js
@@ -15,6 +15,10 @@ import TabsDemo from '@uixkit.react/pages/ComponentsDemo/TabsDemo.js';
 import TabsAnimatedDemo from '@uixkit.react/pages/ComponentsDemo/TabsAnimatedDemo.js';
 import VideoDemo from '@uixkit.react/pages/ComponentsDemo/VideoDemo.js';
 import SwiperDemo from '@uixkit.react/pages/ComponentsDemo/SwiperDemo.js';
+import TableDemo from '@uixkit.react/pages/ComponentsDemo/TableDemo.js';
+import TableGridDemo from '@uixkit.react/pages/ComponentsDemo/TableGridDemo.js';
+import TableSorterDemo from '@uixkit.react/pages/ComponentsDemo/TableSorterDemo.js';
+
 
 
 
@@ -36,6 +40,9 @@ function HookContent() {
 		        <NavLink data-route="true" to={`${url}/tabs-animated`} activeClassName="is-active"><i className="fa fa-check-circle-o" aria-hidden="true"></i> Tabs Animated</NavLink>&nbsp;&nbsp;
 		        <NavLink data-route="true" to={`${url}/video`} activeClassName="is-active"><i className="fa fa-check-circle-o" aria-hidden="true"></i> Video</NavLink>&nbsp;&nbsp;
 		        <NavLink data-route="true" to={`${url}/swiper`} activeClassName="is-active"><i className="fa fa-check-circle-o" aria-hidden="true"></i> Swiper</NavLink>&nbsp;&nbsp;
+		        <NavLink data-route="true" to={`${url}/table`} activeClassName="is-active"><i className="fa fa-check-circle-o" aria-hidden="true"></i> Table</NavLink>&nbsp;&nbsp;
+		        <NavLink data-route="true" to={`${url}/table-grid`} activeClassName="is-active"><i className="fa fa-check-circle-o" aria-hidden="true"></i> Table Grid</NavLink>&nbsp;&nbsp;
+		        <NavLink data-route="true" to={`${url}/table-sorter`} activeClassName="is-active"><i className="fa fa-check-circle-o" aria-hidden="true"></i> Table Sorter</NavLink>&nbsp;&nbsp;
 		    </div>
 
 			 <hr />
@@ -59,6 +66,17 @@ function HookContent() {
 				<Route path={`${path}/swiper`}>
 				  <SwiperDemo />
 				</Route>			
+				<Route path={`${path}/table`}>
+				  <TableDemo />
+				</Route>		
+				<Route path={`${path}/table-grid`}>
+				  <TableGridDemo />
+				</Route>	
+				<Route path={`${path}/table-sorter`}>
+				  <TableSorterDemo />
+				</Route>		
+		
+		
 		
 		
 			  </Switch>